% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{assessment_array_to_flattened_errors}
\alias{assessment_array_to_flattened_errors}
\title{Assessment 3d array to flattened errors}
\usage{
assessment_array_to_flattened_errors(assessments, realizations, error_fun)
}
\arguments{
\item{assessments}{NxExd matrix}

\item{realizations}{N vector}

\item{error_fun}{function taking Mxd matrix and M vector returning Mxd matrix}
}
\value{
Nx(d*E) matrix
}
\description{
Assessment 3d array to flattened errors
}
