% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_copula_posterior.R
\name{calc_pdf_and_cdf_values_flat}
\alias{calc_pdf_and_cdf_values_flat}
\title{Title}
\usage{
calc_pdf_and_cdf_values_flat(eval_points, distributions)
}
\arguments{
\item{eval_points}{nxExd matrix}

\item{distributions}{list of E*d list elements with a cdf and pdf function}
}
\value{
list with two nx(E*d) matrix cdf and pdf values
}
\description{
Title
}
