% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_data.R
\name{combine_expert_lists}
\alias{combine_expert_lists}
\title{Combine the nested formatted data into one list of data frames}
\usage{
combine_expert_lists(formatted_data)
}
\arguments{
\item{formatted_data}{A list of lists of data frames.}
}
\value{
A single list of data frames.
}
\description{
Combine the nested formatted data into one list of data frames
}
