% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/error_metric_class.R
\name{compose_sigmoid_with_error}
\alias{compose_sigmoid_with_error}
\title{Title}
\usage{
compose_sigmoid_with_error(
  error_metric,
  sigmoid,
  sigmoid_prime,
  sigmoid_inverse,
  sigmoid_inverse_prime
)
}
\arguments{
\item{sigmoid}{Should be an increasing function}

\item{sigmoid_inverse_prime}{}
}
\description{
Title
}
