% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/copula_analysis.R
\name{construct_error_distribution_JC_assumption}
\alias{construct_error_distribution_JC_assumption}
\title{Constructs the error distribution from JC assumption. I actually don't think
this is currently needed however.}
\usage{
construct_error_distribution_JC_assumption(
  distribution,
  error_metric,
  expert_m
)
}
\arguments{
\item{distribution}{A single distribution list for one expert}

\item{error_metric}{A error metric class object}

\item{expert_m}{A vector of summarized properties for a single expert. Length d.}
}
\description{
Constructs the error distribution from JC assumption. I actually don't think
this is currently needed however.
}
