% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/copula_analysis.R
\name{construct_error_distributions_JC_assumption}
\alias{construct_error_distributions_JC_assumption}
\title{Title}
\usage{
construct_error_distributions_JC_assumption(
  distributions,
  error_metric,
  assessments
)
}
\arguments{
\item{distributions}{list of n number of distributions}

\item{error_metric}{Error metric of S3 class "error_metric"}

\item{assessments}{nxd matrix or n long vector frame containing the assessments}
}
\value{
List of nxd error distribution. If assessmenets is a nxd matrix then
the output is a nested list with outer length n and inner length d. If
assessments is a n long vector then output is a n long list. Each error
distribution is a list with fields: cdf, pdf and support.
}
\description{
Title
}
