% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_copula_posterior.R
\name{create_log_unnormalized_posterior}
\alias{create_log_unnormalized_posterior}
\title{Title}
\usage{
create_log_unnormalized_posterior(error_copula, error_margins, error_metric, m)
}
\arguments{
\item{error_copula}{Already fitted error copula}

\item{m}{Exd matrix of the summaries properties}
}
\description{
Title
}
