% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_copula_posterior.R
\name{create_log_unnormalized_posterior_JC}
\alias{create_log_unnormalized_posterior_JC}
\title{Returns the log of the unnormalized posterior distribution of the posterior
where the JC assumption of perfect experts is in place.}
\usage{
create_log_unnormalized_posterior_JC(
  error_copula,
  expert_distributions,
  error_metric,
  m
)
}
\arguments{
\item{error_copula}{class "Copula" with dimension d*E. The copula should have its
arguments ordered such that the the first d arguments belongs to expert 1 and
so on.}

\item{expert_distributions}{list of expert distributions. Length E.}

\item{m}{Exd matrix of the summaries properties}
}
\description{
Returns the log of the unnormalized posterior distribution of the posterior
where the JC assumption of perfect experts is in place.
}
