% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_frame_format_utils.R
\name{df_format_to_array_format}
\alias{df_format_to_array_format}
\title{Converts a data frame to a list of arrays suitable for the CopulaSEJ package}
\usage{
df_format_to_array_format(
  training_set,
  test_set,
  m_fun,
  percentiles = c(5, 50, 95)
)
}
\arguments{
\item{training_set, test_set}{data frames with columns expert_id, question_id, and the k_percentiles.}

\item{m_fun}{function to apply to the k_percentiles values to get summary values}

\item{percentiles}{a vector of percentiles 'k' to be used in the data frame.}
}
\value{
a list with three elements:
\itemize{
\item training_summaries: a 3D array of shape (n, E, d) with the training set summaries
\item training_realizations: a vector of length n with the training set realizations
\item test_summaries: a 2D array of shape (E, d) with the test set summaries
}
}
\description{
Converts a data frame to a list of arrays suitable for the CopulaSEJ package
}
