% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/error_metric_class.R
\name{error_to_3d_array}
\alias{error_to_3d_array}
\title{Useful when you want to calculate an error for a fix m matrix but for
multiple different q values.}
\usage{
error_to_3d_array(e_fun, m, q)
}
\arguments{
\item{e_fun}{Takes a axb matrix and a long vector and returns a axb matrix}

\item{m}{Exd long matrix}

\item{q}{N long vector}
}
\value{
a NxExd matrix
}
\description{
Useful when you want to calculate an error for a fix m matrix but for
multiple different q values.
}
