% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/copula_analysis.R
\name{estimate_margin_beta}
\alias{estimate_margin_beta}
\title{Title}
\usage{
estimate_margin_beta(
  obs_vec,
  support = NULL,
  overshoot = 0.1,
  out_of_boundary = "clamp",
  clamp_epsilon = 0.001
)
}
\arguments{
\item{clamp_epsilon}{}
}
\value{
list of pdf, cdf, support, approx_middle. approx middle will be
the mode if the shape parameters are greater than 1. Otherwise it will be
the mean.
}
\description{
Title
}
