% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/copula_analysis.R
\name{estimate_margins}
\alias{estimate_margins}
\title{Estimate marginal distributions}
\usage{
estimate_margins(
  observations,
  supports = NULL,
  method = "beta",
  overshoot = 0.1,
  out_of_boundary = "clamp",
  bw = "SJ",
  beta_mean = 0.5,
  beta_var = 1/12,
  prior_std = 0.1
)
}
\arguments{
\item{observations}{\itemize{
\item a matrix of observations. n rows and d columns
}}
}
\value{
\itemize{
\item a list of d fitted marginal distributions
}
}
\description{
Estimate marginal distributions
}
