% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/copula_analysis.R
\name{find_target_q_support}
\alias{find_target_q_support}
\title{Title}
\usage{
find_target_q_support(
  error_intervals,
  m_test_matrix,
  error_metric,
  combination = "closure"
)
}
\arguments{
\item{error_intervals}{list of c(a,b) intervals}

\item{error_metric}{Error metric of S3 class "error_metric"}

\item{combination}{}
}
\description{
Title
}
