% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/copula_analysis.R
\name{fit_and_construct_posterior}
\alias{fit_and_construct_posterior}
\title{Title}
\usage{
fit_and_construct_posterior(
  training_estimates,
  training_realizations,
  test_matrix,
  copula_model = "joe",
  error_metric = NULL,
  vine_fit_settings = list(),
  error_estimation_settings = list(),
  q_not_cross_zero = TRUE
)
}
\arguments{
\item{training_estimates}{nxExd array with the training data.}

\item{training_realizations}{n long vector with the realisations of the training data.}

\item{test_matrix}{Exd data frame with the test data. Exd order must match the training set.}

\item{copula_model}{string indicating copula model}

\item{error_metric}{Error metric of S3 class "error_metric"}

\item{k_percentiles}{}
}
\description{
Title
}
