% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/copula_analysis.R
\name{fit_and_construct_posterior_indep}
\alias{fit_and_construct_posterior_indep}
\title{Title}
\usage{
fit_and_construct_posterior_indep(
  training_estimates,
  training_realizations,
  test_matrix,
  decoupler,
  copula_model = "joe",
  vine_fit_settings = list(),
  error_estimation_settings = list(),
  q_support_restriction = NULL,
  q_support_overshoot = 0.1,
  rejection_threshold = NULL,
  rejection_min_experts = 3,
  rejection_test = "distance_correlation",
  connection_threshold = NULL
)
}
\arguments{
\item{vine_fit_settings}{List of arguments to pass to the \code{fit_copula} function together with the copula_model argument.}

\item{error_estimation_settings}{List of arguments to pass to the \code{estimate_margins} function}

\item{rejection_test}{}
}
\description{
Title
}
