% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/error_metric_class.R
\name{get_cdf_indep_function_fix_m}
\alias{get_cdf_indep_function_fix_m}
\title{Title}
\usage{
get_cdf_indep_function_fix_m(
  m,
  overshoot = 0.1,
  k_percentiles = c(5, 50, 95),
  scale = "linear",
  support_restriction = NULL
)
}
\arguments{
\item{m}{A matrix ExD}

\item{overshoot}{A scalar}

\item{k_percentiles}{A vector of percentiles}
}
\description{
Title
}
