% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarizing_functions.R
\name{get_median_summarizing_function}
\alias{get_median_summarizing_function}
\title{The median from a set of assessments}
\usage{
get_median_summarizing_function()
}
\description{
The median from a set of assessments
}
