% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interpolate_distributions.R
\name{interpolate_distributions}
\alias{interpolate_distributions}
\title{Constructs a list of distributions for each expert in the data frame. Does
so by linear interpolation of the cdf distribution for each expert.}
\usage{
interpolate_distributions(
  data,
  k_percentiles = c(0, 5, 50, 95, 100),
  interpolation = "linear"
)
}
\arguments{
\item{data}{\itemize{
\item a data frame containing quantile data for each expert. Should
only contain a single question in a single study and the 0 and 100 percentile
must be part of the data.
}}

\item{k_percentiles}{}
}
\value{
a list of one distribution for each expert in ascending order by
expert_id.
}
\description{
Constructs a list of distributions for each expert in the data frame. Does
so by linear interpolation of the cdf distribution for each expert.
}
