% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarizing_functions.R
\name{m_median}
\alias{m_median}
\title{Extracts the median from a set of assessments}
\usage{
m_median(assessments)
}
\arguments{
\item{assessments}{\itemize{
\item Should be a nx3 matrix or data frame
}}
}
\value{
\itemize{
\item An n long vector of the median values of the assessments
}
}
\description{
Extracts the median from a set of assessments
}
