% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarizing_functions.R
\name{m_three_quantiles}
\alias{m_three_quantiles}
\title{Extracts the three quantiles from a set of assessments. Essentially is the
identity function.}
\usage{
m_three_quantiles(assessments)
}
\arguments{
\item{assessments}{\itemize{
\item Should be a nx3 matrix or data frame
}}
}
\value{
\itemize{
\item The same as the input
}
}
\description{
Extracts the three quantiles from a set of assessments. Essentially is the
identity function.
}
