% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{matrix_to_nested_list_by_row}
\alias{matrix_to_nested_list_by_row}
\title{Converts a matrix to a nested list by row}
\usage{
matrix_to_nested_list_by_row(matrix)
}
\arguments{
\item{matrix}{matrix}
}
\value{
A nested list such that list[\link{n}][\link{m}] returns the element in the n-th
row and m-th column of the matrix
}
\description{
Converts a matrix to a nested list by row
}
