% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/error_metric_class.R
\name{new_decoupler}
\alias{new_decoupler}
\title{Title}
\usage{
new_decoupler(
  name,
  f,
  f_prime_q,
  f_increasing,
  D_tilde,
  fix_m = NULL,
  f_inverse = NULL,
  short_name = NULL,
  ideal_mean_var = NULL
)
}
\arguments{
\item{f, f_prime_q}{(q,m) returns (NxExD)}

\item{ideal_mean_var}{}
}
\description{
Title
}
