% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distribution_class.R
\name{new_distribution}
\alias{new_distribution}
\title{Distribution object}
\usage{
new_distribution(cdf, pdf, support, sample = NULL, cdf_inv = NULL)
}
\arguments{
\item{cdf, pdf, cdf_inv}{of the distribution}

\item{support}{of the distribution. An interval \link{a,b} such that the pdf is 0
outside of \link{a,b}. Optional}

\item{sample}{function to sample from the distribution. Optional.}
}
\value{
A new distribution object with class distribution
}
\description{
Distribution object
}
