% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/error_metric_class.R
\name{new_error_metric}
\alias{new_error_metric}
\title{Constructor for error metric class}
\usage{
new_error_metric(
  name,
  f,
  f_prime_m,
  f_prime_inverse_q,
  f_increasing_q,
  f_inverse_q = NULL
)
}
\arguments{
\item{name}{name of the function}

\item{f, f_prime_m}{The error metric and the derivative of f with respect to
m. Both functions have two arguments (m,q) where m is a nxd matrix and
q is a vector of length n. Returns a matrix of the same shape as m.}

\item{f_increasing_q}{Function of a single argument (m). f_inverse_q is
optional.
Outputs a Boolean vector or matrix in the same shape as m. True if f is
increasing in q, False if not, NA if not defined.}

\item{f_inverse_q, f_prime_inverse_q}{The inverse of f with respect to
q and the derivative of the inverse respectively. Have two arguments (m,e)
where m and e are vectors with length d and L. Outputs a matrix of size Lxd}
}
\value{
A new error metric object
}
\description{
Constructor for error metric class
}
