% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_frame_format_utils.R
\name{percentiles_from_dataframe}
\alias{percentiles_from_dataframe}
\title{Extracts the percentile columns from the dataframe and returns a matrix}
\usage{
percentiles_from_dataframe(df, percentiles)
}
\arguments{
\item{percentiles}{\itemize{
\item the percentiles present in the data frame. Length d. Integer
}}

\item{study_data}{\itemize{
\item a data frame with percentile columns. n rows
}}
}
\value{
\itemize{
\item a matrix (nxd) with the assessment data
}
}
\description{
Extracts the percentile columns from the dataframe and returns a matrix
}
