% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/posterior_estimates.R
\name{sample_log_unnormalized_density}
\alias{sample_log_unnormalized_density}
\title{Title}
\usage{
sample_log_unnormalized_density(
  log_density,
  support,
  num_samples,
  start_point = NULL,
  method = "BayesianTools"
)
}
\arguments{
\item{start_point}{Needs to be provided if the support is infinite}
}
\description{
Title
}
