% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{split_dataframe_to_error_observations}
\alias{split_dataframe_to_error_observations}
\title{Splits data frame by questions and structures error order}
\usage{
split_dataframe_to_error_observations(
  dataframe,
  error_metric,
  summarizing_function,
  k_percentiles = c(5, 50, 95)
)
}
\arguments{
\item{dataframe}{frame data frame with columns expert_id, question_id, and the k_percentiles}

\item{error_metric}{error metric object}

\item{summarizing_function}{summarizing function}

\item{k_percentiles}{vector of percentiles}
}
\value{
n x (d*E) matrix
}
\description{
Takes a data frame with the columns expert_id, question_id, and the k_percentiles
and splits the data frame by question_id (in ascending order). For each sub
data frame the errors are calculated and put into a d\emph{E long vector ordered
such that the errors from a single expert are kept adjacent. These vectors are then
put row by row in a matrix resulting in a nx(d}E) matrix where n is the
number of questions.
}
