% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/copula_analysis.R
\name{study_test_performance}
\alias{study_test_performance}
\title{Title}
\usage{
study_test_performance(study_data, sim_params = NULL)
}
\arguments{
\item{study_data}{\itemize{
\item a data frame with columns expert_id, question_id, and the
k_percentiles, realization
}}

\item{k_percentiles}{}
}
\description{
Title
}
