% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{support_intersection}
\alias{support_intersection}
\title{Calculates the intersection of a list of supports.}
\usage{
support_intersection(support_list, nested_list = FALSE)
}
\arguments{
\item{support_list, nested_list}{a named list of elements with a 'support'
element being a numeric vector of length 2 if nested_list=TRUE. Otherwise
directly a list with the numeric vectors as elements.}
}
\description{
Throws warning if the support is 0 wide
}
