% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/error_metric_class.R
\name{compose_sigmoid_with_decoupler}
\alias{compose_sigmoid_with_decoupler}
\title{Title}
\usage{
compose_sigmoid_with_decoupler(
  decoupler,
  sigmoid_f,
  sigmoid_prime_f,
  sigmoid_inverse_f
)
}
\arguments{
\item{sigmoid}{Should be an increasing function}

\item{sigmoid_inverse}{}
}
\description{
Title
}
