% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarizing_functions.R
\name{new_summarizing_function}
\alias{new_summarizing_function}
\title{A class to hold a summarizing function}
\usage{
new_summarizing_function(f, name = "", output_dim = NULL, short_name = "")
}
\arguments{
\item{f}{Function that takes a nx3 matrix of assessments and returns a nxd matrix}

\item{name}{Descriptive name of summarizing function}

\item{output_dim}{The d dimension of the summarizing function}

\item{short_name}{Short name of the summarizing function. Should have no spaces and be filesystem friendly.}
}
\description{
A class to hold a summarizing function
}
